Working, but dirty code. Sloppiness abounds, but it is fully functional. Next step would be to just make it cleaner for milestone, and then figure out how to 'run' it better (possibly a 'test' via bus as per design)

+++ Practicum-Github-GizmoE/src/gizmoe/taskexecutor/CapabilityMessageListener.java
@@ -19,7 +19,7 @@ public class CapabilityMessageListener implements MessageListener{
 					ConcurrentHashMap<String, Object> output = (ConcurrentHashMap<String, Object>) inMsg.getObject();
                 	//int toPrint = (Integer) output.get("out");
                 	for(Object toPrint : output.values()){
-                		System.out.println("TaskExecutor:: "+queue+" gave me:: "+(Integer)toPrint);
+                		System.out.println("TaskExecutor:: "+queue+" gave me:: "+toPrint);
                 	}
                 	TaskExecutor.registerOutputs(id, output);
                 }
